benchmarks/kn/original/video-analytics/video-streaming/requirements.txt
parliament-functions==0.1.0
requests
